adds Uncle Bob methodology
This code refactor aligns the design closer to the one that professor coleman has discussed with us. Furthermore, it bring the code in line with uncle Bob's idea of no function having more than 4 lines of code. Currently only three functions are in violation of this rule. Two of them have 7 lines of code and the other has 13 lines of code.